Try compile with mingw.
