+++ encoding_lib/CMakeLists.txt
@@ -1,5 +1,11 @@
 project(arithmetic-encoding-lib)
 
+set(Boost_USE_STATIC_LIBS ON)
+set(Boost_USE_MULTITHREADED ON)
+set(Boost_USE_STATIC_RUNTIME OFF)
+
+find_package(Boost COMPONENTS system program_options REQUIRED)
+
 add_library(${PROJECT_NAME} STATIC)
 
 target_sources(${PROJECT_NAME}
@@ -10,12 +16,15 @@ target_sources(${PROJECT_NAME}
         src/byte_data_constructor.cpp
         src/data_parser.cpp
         src/decreasing_on_update_dictionary.cpp        
-        src/ppmd_dictionary.cpp
         src/ppma_dictionary.cpp
+        src/ppmd_dictionary.cpp
+        src/ranges_calc.cpp
         src/static_dictionary.cpp
         src/uniform_dictionary.cpp
         src/integer_random_access_iterator.hpp
 )
+
+include_directories(${Boost_INCLUDE_DIRS})
 
 target_include_directories(${PROJECT_NAME}
     PRIVATE
@@ -37,7 +46,7 @@ set(public_headers
     include/ael/bits_iterator.hpp
     include/ael/byte_data_constructor.hpp
     include/ael/data_parser.hpp
-    include/ael/ranges_calc.hpp
+    include/ael/impl/ranges_calc.hpp
     include/ael/dictionary/impl/a_d_dictionary_base.hpp
     include/ael/dictionary/uniform_dictionary.hpp
     include/ael/dictionary/static_dictionary.hpp
@@ -62,3 +71,7 @@ set_target_properties(${PROJECT_NAME} PROPERTIES PUBLIC_HEADER "${public_headers
 set_target_properties(${PROJECT_NAME} PROPERTIES LINKER_LANGUAGE CXX)
 
 target_link_libraries(arithmetic-encoding-lib dynamic-segment-tree)
+
+if (BUILD_TEST)
+    add_subdirectory(test)
+endif (BUILD_TEST)
