+++ encoding_lib/src/ranges_calc.cpp
@@ -1,48 +1,6 @@
-#ifndef RANGES_CALC_HPP
-#define RANGES_CALC_HPP
+#include <ael/impl/ranges_calc.hpp>
 
-#include <cstdint>
-#include <iostream>
-#include <boost/multiprecision/cpp_int.hpp>
-
-namespace ael {
-
-namespace bm = boost::multiprecision;
-
-////////////////////////////////////////////////////////////////////////////////
-/// \brief The RangesCalc class
-///
-class RangesCalc {
-public:
-
-    using Count = std::uint64_t;
-    struct Range;
-
-public:
-
-    constexpr static const std::uint16_t numBits = 56;
-    constexpr static const Count total = Count{1} << Count{numBits};
-    constexpr static const Count half = Count{1} << Count{numBits - 1};
-    constexpr static const Count quater = Count{1} << Count{numBits - 2};
-    constexpr static const Count threeQuaters = 3 * quater;
-
-public:
-
-    static Range recalcRange(Range r);
-
-    static Range rangeFromStatsAndPrev(
-            Range r, Count low, Count high, Count total);
-
-};
-
-////////////////////////////////////////////////////////////////////////////////
-/// \brief The Range class
-///
-struct RangesCalc::Range {
-    Count low;
-    Count high;
-};
-
+namespace ael::impl {
 ////////////////////////////////////////////////////////////////////////////////
 //----------------------------------------------------------------------------//
 auto RangesCalc::recalcRange(Range r) -> Range {
@@ -73,7 +31,4 @@ auto RangesCalc::rangeFromStatsAndPrev(
     };
 }
 
-
 }
-
-#endif // RANGES_CALC_HPP
